Moved tests into their own folder
