+++ __tests__/components/App/App.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { MemoryRouter } from 'react-router-dom';
 
-import App from './App';
+import App from '../../../src/components/App';
 import { provideState } from '../../testHelpers';
 
 describe('<App />', () => {
