+++ package.json
@@ -33,7 +33,7 @@
   "scripts": {
     "start": "concurrently -r \"yarn run start-react\" \"wait-on http://localhost:3000/ && yarn run start-electron\" --kill-others --success first",
     "build": "react-app-rewired build",
-    "test": "react-app-rewired test --env=jsdom",
+    "test": "react-app-rewired test --env=jsdom --testMatch '**/__tests__/**/*.test.{js,jsx,mjs}' --setupTestFrameworkScriptFile './__tests__/setupTests.js'",
     "start-react": "cross-env BROWSER=none react-app-rewired start",
     "start-electron": "cross-env ELECTRON_START_URL=http://localhost:3000/ electron .",
     "pack": "yarn run build && electron-builder --dir",
